backend/requirements.txt
flask
flask-mongoengine
pymongo[srv]
PyGithub
PyQuery
python-dotenv
flask-cors